+++ back-end/src/main/java/index/PsqlSpatialIndexer.java
@@ -18,19 +18,14 @@ import project.Transform;
 public class PsqlSpatialIndexer extends BoundingBoxIndexer {
 
     private static PsqlSpatialIndexer instance = null;
-    private static boolean isCitus = false;
 
     // singleton pattern to ensure only one instance existed
     protected PsqlSpatialIndexer() {}
 
-    private PsqlSpatialIndexer(boolean isCitus) {
-        this.isCitus = isCitus;
-    }
-
     // thread-safe instance getter
-    public static synchronized PsqlSpatialIndexer getInstance(boolean isCitus) {
+    public static synchronized PsqlSpatialIndexer getInstance() {
 
-        if (instance == null) instance = new PsqlSpatialIndexer(isCitus);
+        if (instance == null) instance = new PsqlSpatialIndexer();
         return instance;
     }
 
@@ -69,22 +64,10 @@ public class PsqlSpatialIndexer extends BoundingBoxIndexer {
         sql = "create table " + bboxTableName + " (";
         for (int i = 0; i < trans.getColumnNames().size(); i++)
             sql += trans.getColumnNames().get(i) + " text, ";
-        if (isCitus) {
-            sql += "citus_distribution_id int, ";
-        }
         sql +=
                 "cx double precision, cy double precision, minx double precision, miny double precision, maxx double precision, maxy double precision, geom geometry(polygon));";
         System.out.println(sql);
         bboxStmt.executeUpdate(sql);
-
-        if (isCitus) {
-            sql =
-                    "SELECT create_distributed_table('"
-                            + bboxTableName
-                            + "', 'citus_distribution_id');";
-            System.out.println(sql);
-            bboxStmt.executeQuery(sql);
-        }
 
         // if this is an empty layer, return
         if (trans.getDb().isEmpty()) return;
@@ -97,9 +80,6 @@ public class PsqlSpatialIndexer extends BoundingBoxIndexer {
         // TODO: distinguish between separable and non-separable cases
         String insertSql = "insert into " + bboxTableName + " values (";
         for (int i = 0; i < trans.getColumnNames().size() + 6; i++) insertSql += "?, ";
-        if (isCitus) {
-            insertSql += "?, ";
-        }
         insertSql += "ST_GeomFromText(?));";
 
         System.out.println(insertSql);
@@ -147,10 +127,6 @@ public class PsqlSpatialIndexer extends BoundingBoxIndexer {
             int pscol = 1;
             for (int i = 0; i < transformedRow.size(); i++)
                 preparedStmt.setString(pscol++, transformedRow.get(i).replaceAll("\'", "\'\'"));
-            if (isCitus) {
-                // row number is a fine distribution key (for now) - round robin across the cluster
-                preparedStmt.setInt(pscol++, rowCount);
-            }
             for (int i = 0; i < 6; i++) preparedStmt.setDouble(pscol++, curBbox.get(i));
 
             double minx, miny, maxx, maxy;
