+++ back-end/src/main/java/index/PsqlTileIndexer.java
@@ -18,17 +18,14 @@ import project.Transform;
 public class PsqlTileIndexer extends BoundingBoxIndexer {
 
     private static PsqlTileIndexer instance = null;
-    private static boolean isCitus = false;
 
     // singleton pattern to ensure only one instance existed
-    private PsqlTileIndexer(boolean isCitus) {
-        this.isCitus = isCitus;
-    }
+    private PsqlTileIndexer() {}
 
     // thread-safe instance getter
-    public static synchronized PsqlTileIndexer getInstance(boolean isCitus) {
+    public static synchronized PsqlTileIndexer getInstance() {
 
-        if (instance == null) instance = new PsqlTileIndexer(isCitus);
+        if (instance == null) instance = new PsqlTileIndexer();
         return instance;
     }
 
@@ -69,9 +66,6 @@ public class PsqlTileIndexer extends BoundingBoxIndexer {
         sql = "create table " + bboxTableName + " (";
         for (int i = 0; i < trans.getColumnNames().size(); i++)
             sql += trans.getColumnNames().get(i) + " text, ";
-        if (isCitus) {
-            sql += "citus_distribution_id int, ";
-        }
         sql +=
                 "tuple_id int, cx double precision, cy double precision, minx double "
                         + "precision, miny double precision, maxx double precision, maxy double precision);";
@@ -83,7 +77,6 @@ public class PsqlTileIndexer extends BoundingBoxIndexer {
                         + tileTableName
                         + " ("
                         + "tuple_id int, "
-                        + (isCitus ? "citus_distribution_id int, " : "")
                         + "tile_id varchar(50)"
                         + ");";
         tileStmt.executeUpdate(sql);
@@ -130,10 +123,6 @@ public class PsqlTileIndexer extends BoundingBoxIndexer {
             for (int i = 0; i < transformedRow.size(); i++)
                 bboxInsSqlBuilder.append(
                         "'" + transformedRow.get(i).replaceAll("\'", "\'\'") + "', ");
-            if (isCitus) {
-                // row number is a fine distribution key (for now) - round robin across the cluster
-                bboxInsSqlBuilder.append(rowCount);
-            }
             bboxInsSqlBuilder.append(String.valueOf(rowCount));
             for (int i = 0; i < 6; i++)
                 bboxInsSqlBuilder.append(", " + String.valueOf(curBbox.get(i)));
@@ -163,16 +152,7 @@ public class PsqlTileIndexer extends BoundingBoxIndexer {
                         if (tileInsSqlBuilder.charAt(tileInsSqlBuilder.length() - 1) == ')')
                             tileInsSqlBuilder.append(",(");
                         else tileInsSqlBuilder.append(" (");
-                        tileInsSqlBuilder.append(
-                                rowCount
-                                        + ", "
-                                        +
-                                        // rownum is a fine distrib key (for now) - round robin
-                                        // across the cluster
-                                        (isCitus ? (rowCount + ", ") : "")
-                                        + "'"
-                                        + tileId
-                                        + "')");
+                        tileInsSqlBuilder.append(rowCount + ", '" + tileId + "')");
                         if (mappingCount % Config.tileBatchSize == 0) {
                             tileInsSqlBuilder.append(";");
                             tileStmt.executeUpdate(tileInsSqlBuilder.toString());
